# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Books API",
        "description": "API Documentation for Books",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Returns list all book",
                "description": "API for getting all books data",
                "operationId": "booksID",
                "responses": {
                    "default": {
                        "description": "succesfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book added successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/book/store": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "adding book to database",
                "description": "adding book to database",
                "operationId": "addBookID",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "penulis",
                                    "harga",
                                    "tanggal_terbit"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "maxLength": 60
                                    },
                                    "penulis": {
                                        "type": "string",
                                        "maxLength": 60
                                    },
                                    "harga": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "tanggal_terbit": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "sampul_buku": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book added successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookResource": {
                "title": "Book Resource",
                "description": "Schema for book details",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "judul": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "penulis": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                    },
                    "harga": {
                        "type": "number",
                        "format": "float",
                        "example": 10.99
                    },
                    "tanggal_terbit": {
                        "type": "string",
                        "format": "date",
                        "example": "1925-04-10"
                    },
                    "sampul_buku_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/sampul.jpg"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "books",
            "description": "books"
        }
    ]
}